+++ packages/react-tinacms-inline/src/styles/focus-ring.tsx
@@ -38,8 +38,8 @@ export const FocusRing = styled.div<FocusRingProps>(
       content: '';
       display: block;
       position: absolute;
-      left: calc(-1 * ${p.offset ? p.offset : '16'}px);
-      top: calc(-1 * ${p.offset ? p.offset : '16'}px);
+      left: calc(-1 * ${p.offset !== undefined ? p.offset : '16'}px);
+      top: calc(-1 * ${p.offset !== undefined ? p.offset : '16'}px);
       width: calc(100% + ${p.offset !== undefined ? p.offset * 2 : '32'}px);
       height: calc(100% + ${p.offset !== undefined ? p.offset * 2 : '32'}px);
       border: 1px solid var(--tina-color-primary);
@@ -47,10 +47,12 @@ export const FocusRing = styled.div<FocusRingProps>(
       opacity: 0;
       pointer-events: none;
       transition: all var(--tina-timing-medium) ease-out;
+      box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
     }
 
     ${p.active &&
       css`
+        &:hover:after,
         &:after {
           opacity: 1;
         }