+++ packages/next-tinacms-markdown/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.18-alpha.1](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.18-alpha.0...next-tinacms-markdown@0.2.18-alpha.1) (2020-05-29)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
 ## [0.2.18-alpha.0](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.17...next-tinacms-markdown@0.2.18-alpha.0) (2020-05-28)
 
 **Note:** Version bump only for package next-tinacms-markdown