+++ packages/react-tinacms-github/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1-alpha.2](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.4.1-alpha.1...react-tinacms-github@0.4.1-alpha.2) (2020-05-29)
+
+
+### Bug Fixes
+
+* set sha for new files ([fc06337](https://github.com/tinacms/tinacms/commit/fc06337)), closes [#1172](https://github.com/tinacms/tinacms/issues/1172)
+
+
+
+
+
 ## [0.4.1-alpha.1](https://github.com/tinacms/tinacms/compare/react-tinacms-github@0.4.1-alpha.0...react-tinacms-github@0.4.1-alpha.1) (2020-05-28)